build(amazonq): remove standalone local context server (#5654)
funtionality is already being provided through Flare
