+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonq/lsp/artifacts/ArtifactHelper.kt
@@ -17,7 +17,6 @@ import software.aws.toolkits.core.utils.getLogger
 import software.aws.toolkits.core.utils.info
 import software.aws.toolkits.core.utils.warn
 import software.aws.toolkits.jetbrains.core.saveFileFromUrl
-import software.aws.toolkits.jetbrains.services.amazonq.project.manifest.ManifestManager
 import software.aws.toolkits.resources.AwsCoreBundle
 import java.nio.file.Files
 import java.nio.file.Path
@@ -29,7 +28,7 @@ class ArtifactHelper internal constructor(
 ) {
     private val currentAttempt = AtomicInteger(0)
 
-    fun removeDelistedVersions(delistedVersions: List<ManifestManager.Version>) {
+    fun removeDelistedVersions(delistedVersions: List<Version>) {
         val localFolders = getSubFolders(lspArtifactsPath)
 
         delistedVersions.forEach { delistedVersion ->
@@ -78,7 +77,7 @@ class ArtifactHelper internal constructor(
             .sortedByDescending { (_, semVer) -> semVer }
     }
 
-    fun getExistingLspArtifacts(versions: List<ManifestManager.Version>, target: ManifestManager.VersionTarget?): Boolean {
+    fun getExistingLspArtifacts(versions: List<Version>, target: VersionTarget?): Boolean {
         if (versions.isEmpty() || target?.contents == null) return false
 
         val localLSPPath = lspArtifactsPath.resolve(versions.first().serverVersion.toString())
@@ -102,7 +101,7 @@ class ArtifactHelper internal constructor(
         return !hasInvalidFiles
     }
 
-    suspend fun tryDownloadLspArtifacts(project: Project, versions: List<ManifestManager.Version>, target: ManifestManager.VersionTarget?): Path? {
+    suspend fun tryDownloadLspArtifacts(project: Project, versions: List<Version>, target: VersionTarget?): Path? {
         val temporaryDownloadPath = Files.createTempDirectory("lsp-dl")
         val downloadPath = lspArtifactsPath.resolve(versions.first().serverVersion.toString())
 
@@ -145,7 +144,7 @@ class ArtifactHelper internal constructor(
     }
 
     @VisibleForTesting
-    internal fun downloadLspArtifacts(downloadPath: Path, target: ManifestManager.VersionTarget?): Boolean {
+    internal fun downloadLspArtifacts(downloadPath: Path, target: VersionTarget?): Boolean {
         if (target == null || target.contents.isNullOrEmpty()) {
             logger.warn { "No target contents available for download" }
             return false
@@ -199,7 +198,7 @@ class ArtifactHelper internal constructor(
         return "sha384:$contentHash" == expectedHash
     }
 
-    private fun validateDownloadedFiles(downloadPath: Path, contents: List<ManifestManager.TargetContent>) {
+    private fun validateDownloadedFiles(downloadPath: Path, contents: List<TargetContent>) {
         val missingFiles = contents
             .mapNotNull { it.filename }
             .filter { filename ->
