+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonq/lsp/artifacts/ArtifactManager.kt
@@ -16,7 +16,6 @@ import org.jetbrains.annotations.VisibleForTesting
 import software.aws.toolkits.core.utils.error
 import software.aws.toolkits.core.utils.getLogger
 import software.aws.toolkits.core.utils.info
-import software.aws.toolkits.jetbrains.services.amazonq.project.manifest.ManifestManager
 import java.nio.file.Path
 
 @Service
@@ -35,8 +34,8 @@ class ArtifactManager @NonInjectable internal constructor(private val manifestFe
         val endVersion: SemVer,
     )
     data class LSPVersions(
-        val deListedVersions: List<ManifestManager.Version>,
-        val inRangeVersions: List<ManifestManager.Version>,
+        val deListedVersions: List<Version>,
+        val inRangeVersions: List<Version>,
     )
 
     companion object {
@@ -91,7 +90,7 @@ class ArtifactManager @NonInjectable internal constructor(private val manifestFe
     }
 
     @VisibleForTesting
-    internal fun getLSPVersionsFromManifestWithSpecifiedRange(manifest: ManifestManager.Manifest): LSPVersions {
+    internal fun getLSPVersionsFromManifestWithSpecifiedRange(manifest: Manifest): LSPVersions {
         if (manifest.versions.isNullOrEmpty()) return LSPVersions(emptyList(), emptyList())
 
         val (deListed, inRange) = manifest.versions.mapNotNull { version ->
@@ -112,7 +111,7 @@ class ArtifactManager @NonInjectable internal constructor(private val manifestFe
         )
     }
 
-    private fun getTargetFromLspManifest(versions: List<ManifestManager.Version>): ManifestManager.VersionTarget {
+    private fun getTargetFromLspManifest(versions: List<Version>): VersionTarget {
         val currentOS = getCurrentOS()
         val currentArchitecture = getCurrentArchitecture()
