+++ plugins/amazonq/shared/jetbrains-community/tst/software/aws/toolkits/jetbrains/services/amazonq/lsp/artifacts/ArtifactManagerTest.kt
@@ -21,7 +21,6 @@ import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.api.extension.RegisterExtension
 import org.junit.jupiter.api.io.TempDir
 import software.aws.toolkits.jetbrains.services.amazonq.lsp.artifacts.ArtifactManager.SupportedManifestVersionRange
-import software.aws.toolkits.jetbrains.services.amazonq.project.manifest.ManifestManager
 import java.nio.file.Path
 
 class ArtifactManagerTest {
@@ -64,7 +63,7 @@ class ArtifactManagerTest {
 
     @Test
     fun `fetch artifact does not have any valid lsp versions`() = runTest {
-        every { manifestFetcher.fetch() }.returns(ManifestManager.Manifest())
+        every { manifestFetcher.fetch() }.returns(Manifest())
 
         every { artifactManager.getLSPVersionsFromManifestWithSpecifiedRange(any()) }.returns(
             ArtifactManager.LSPVersions(deListedVersions = emptyList(), inRangeVersions = emptyList())
@@ -81,7 +80,7 @@ class ArtifactManagerTest {
     fun `fetch artifact if inRangeVersions are not available should fallback to local lsp`() = runTest {
         val expectedResult = listOf(Pair(tempDir, SemVer("1.0.0", 1, 0, 0)))
 
-        every { manifestFetcher.fetch() }.returns(ManifestManager.Manifest())
+        every { manifestFetcher.fetch() }.returns(Manifest())
         every { artifactHelper.getAllLocalLspArtifactsWithinManifestRange(any()) }.returns(expectedResult)
 
         artifactManager.fetchArtifact(projectExtension.project)
@@ -92,13 +91,13 @@ class ArtifactManagerTest {
 
     @Test
     fun `fetch artifact have valid version in local system`() = runTest {
-        val target = ManifestManager.VersionTarget(platform = "temp", arch = "temp")
-        val versions = listOf(ManifestManager.Version("1.0.0", targets = listOf(target)))
+        val target = VersionTarget(platform = "temp", arch = "temp")
+        val versions = listOf(Version("1.0.0", targets = listOf(target)))
 
         every { artifactManager.getLSPVersionsFromManifestWithSpecifiedRange(any()) }.returns(
             ArtifactManager.LSPVersions(deListedVersions = emptyList(), inRangeVersions = versions)
         )
-        every { manifestFetcher.fetch() }.returns(ManifestManager.Manifest())
+        every { manifestFetcher.fetch() }.returns(Manifest())
 
         mockkStatic("software.aws.toolkits.jetbrains.services.amazonq.lsp.artifacts.LspUtilsKt")
         every { getCurrentOS() }.returns("temp")
@@ -116,14 +115,14 @@ class ArtifactManagerTest {
 
     @Test
     fun `fetch artifact does not have valid version in local system`() = runTest {
-        val target = ManifestManager.VersionTarget(platform = "temp", arch = "temp")
-        val versions = listOf(ManifestManager.Version("1.0.0", targets = listOf(target)))
+        val target = VersionTarget(platform = "temp", arch = "temp")
+        val versions = listOf(Version("1.0.0", targets = listOf(target)))
         val expectedResult = listOf(Pair(tempDir, SemVer("1.0.0", 1, 0, 0)))
 
         every { artifactManager.getLSPVersionsFromManifestWithSpecifiedRange(any()) }.returns(
             ArtifactManager.LSPVersions(deListedVersions = emptyList(), inRangeVersions = versions)
         )
-        every { manifestFetcher.fetch() }.returns(ManifestManager.Manifest())
+        every { manifestFetcher.fetch() }.returns(Manifest())
 
         mockkStatic("software.aws.toolkits.jetbrains.services.amazonq.lsp.artifacts.LspUtilsKt")
         every { getCurrentOS() }.returns("temp")
